Functionality now in line with the original
The errors appear to have been corrected. There are still possible
issues that need to be changed, and it comes down to issues that were
present in the original. But the code is now much more condensed, and
the functions for handling each player is now generic, meaning they will
more easily be amenable to modeling.