Random Behavior, New Functions, More Cleanup
A random agent has been installed (see "aRandomTurn()"). By clicking
anywhere (though preferably not on the button), each tenth of a second,
an agent will randomly choose an action from the available pull.

Functions have been added to facilitate this, and for the implementation
of the intelligent agent afterward (see lines 311 - 367).

The function to check if a double is blocking a player's movement
(isBlocked()) is currently bugged and is not included in the logic for
valid moves. Will be fixed in the next update.

Next update will include the ability to choose the configuration of
agents (how many humans, random & rational agents).