Fixed the block check (and other improvements)
Checking for if a player is blocked should now work as intended.

Before the game starts, the player is asked how many humans they want to
play, and then if there are any computers, the split of random vs.
intelligent.

The dialog boxes are a bit finicky--they show up at different places and
occasionally hide the game window; the reason for this behavior is
unknown.

When a player wins, that information is now displayed on the game board
instead of in the console.