Wrote getNextStatesForActions() (and other changes)
The aforementioned function has been written in, but the behavior has
been commented out since it has to be in the ludo_game module to work.
The necessary change of removing the parameter of stateNum in minimax,
as well as other changes that have to be made to the function inside,
have been added.
ludo_game.py is still a work in progress.